generalised transformation1
